Moved default gamemode placeholders enum to their own file
